requirement.txt
ccxt 
python-binance
finta